Agregar homepage con redirección al panel y corrección del enlace de retorno
